feat: loadable, triggers, playbook
