+++ index.html
@@ -6,25 +6,11 @@
       rel="icon"
       type="image/svg+xml"
       href="/images/favicon.svg"
-      media="(prefers-color-scheme: light)"
-    />
-    <link
-      rel="icon"
-      type="image/svg+xml"
-      href="/images/favicon_light.svg"
-      media="(prefers-color-scheme: dark)"
     />
     <link
       rel="icon"
       type="image/png"
       href="/images/favicon.png"
-      media="(prefers-color-scheme: light)"
-    />
-    <link
-      rel="icon"
-      type="image/png"
-      href="/images/favicon_light.png"
-      media="(prefers-color-scheme: dark)"
     />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
