+++ src/components/country-selector.tsx
@@ -1,18 +1,35 @@
 import { type Control, Controller } from 'react-hook-form';
-import { MultiSelect, type MultiSelectOption } from '@/components/ui/multi-select';
-import { COUNTRIES } from '@/features/icp/constants/countries';
+import { MultiSelect, type MultiSelectGroup, type MultiSelectOption } from '@/components/ui/multi-select';
+import { COUNTRIES, REGIONS } from '@/constants/countries';
 
 interface CountrySelectorProps {
   control: Control<any>;
   name: string;
+  withRegions?: boolean;
 }
 
-export default function CountrySelector({ control, name }: CountrySelectorProps) {
-  // Convert countries to MultiSelectOption format
-  const countryOptions: MultiSelectOption[] = COUNTRIES.map(country => ({
-    label: country.label,
-    value: country.value,
-  }));
+export default function CountrySelector({ control, name, withRegions = false }: CountrySelectorProps) {
+  const options: MultiSelectOption[] | MultiSelectGroup[] = withRegions 
+    ? [
+        {
+          heading: "Regions",
+          options: REGIONS.map(region => ({
+            label: region.label,
+            value: region.value,
+          })),
+        },
+        {
+          heading: "Countries", 
+          options: COUNTRIES.map(country => ({
+            label: country.label,
+            value: country.value,
+          })),
+        },
+      ]
+    : COUNTRIES.map(country => ({
+        label: country.label,
+        value: country.value,
+      }));
 
   return (
     <Controller
@@ -21,10 +38,12 @@ export default function CountrySelector({ control, name }: CountrySelectorProps)
       defaultValue={[]}
       render={({ field }) => (
         <MultiSelect
-          options={countryOptions}
+          options={options}
+          maxCount={6}
+          animationConfig={{ optionHoverAnimation: 'none', badgeAnimation: 'none' }}
           onValueChange={field.onChange}
           defaultValue={field.value || []}
-          placeholder="Select countries..."
+          placeholder={withRegions ? "Select countries or regions..." : "Select countries..."}
           searchable={true}
           className="w-full"
         />
