--- src/routes/playbooks/$playbookId.tsx
@@ -1,6 +1,0 @@
-import { createFileRoute } from '@tanstack/react-router';
-import PlaybookDetailPage from '@/features/playbooks/detail';
-
-export const Route = createFileRoute('/playbooks/$playbookId')({
-  component: PlaybookDetailPage,
-});
