--- src/routes/playbooks/index.tsx
@@ -1,5 +1,0 @@
-import { createFileRoute } from '@tanstack/react-router';
-import PlaybooksPage from '@/features/playbooks';
-
-export const Route = createFileRoute('/playbooks/')({  component: PlaybooksPage,
-});
