+++ src/routes/plays/index.tsx
@@ -1,0 +1,5 @@
+import { createFileRoute } from '@tanstack/react-router';
+import PlaybooksPage from '@/features/plays';
+
+export const Route = createFileRoute('/plays/')({  component: PlaybooksPage,
+});
