+++ src/routes/plays/new.tsx
@@ -1,0 +1,6 @@
+import { createFileRoute } from '@tanstack/react-router';
+import PlaybookNewPage from '@/features/plays/new';
+
+export const Route = createFileRoute('/plays/new')({
+  component: PlaybookNewPage,
+});
